Revert to local log until a better cloud solution is determined
If file based it is as easy as uploading the data stored locally to the cloud.
